It seems to me that the "Upload" button was not recognized by the Forms API because it was only generated inside an element process function.
Fix that by going back to a more imagefield like approach, but not quite -
upload and editing widgets are now separated, with the upload widget only
returning the fid. The combined form element is now gone, and the widget
is directly composed in filefield_widget().

Bonus feature (on the way to deprecating imagefield): other modules can now
provide their own replacement edit forms, with priority based ranking of
which one is used. Might need a bit of work, but it's a good start.